[PNYM-13] Implemented Room database and bound it with TransactionsRepositoryCoreImpl.kt
DI integration

Updated Transaction and Category: replaced Int id with Long

+++ core/ui/navigation/src/main/java/io/github/t1geryan/navigation/RootNavEntry.kt
@@ -9,10 +9,10 @@ sealed interface RootNavEntry {
     data object Tabs
 
     @Serializable
-    data class CreateOrUpdateTransaction(val id: Int?)
+    data class CreateOrUpdateTransaction(val id: Long?)
 
     @Serializable
-    data class CreateOrUpdateCategory(val id: Int?)
+    data class CreateOrUpdateCategory(val id: Long?)
 
     companion object {
         val INITIAL = Tabs
