+++ domain/src/main/java/io/github/t1geryan/domain/models/Transaction.kt
@@ -3,11 +3,11 @@ package io.github.t1geryan.domain.models
 import androidx.compose.runtime.Immutable
 import kotlinx.datetime.LocalDateTime
 
-typealias TransactionId = Int
+typealias TransactionId = Long
 
 @Immutable
 data class Transaction(
-    val id: TransactionId = 0,
+    val id: TransactionId = 0L,
     val name: String,
     val amount: Amount,
     val category: Category,
